crew-ai-storybook/requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>= 0.0.2
python-dotenv>=1.0.0
